ref(amw_rest): move new endpoints to dedicated class
