+++ AMW_rest/src/test/java/ch/mobi/itc/mobiliar/rest/resources/ResourceGroupsRestTest.java
@@ -20,39 +20,12 @@
 
 package ch.mobi.itc.mobiliar.rest.resources;
 
-import static ch.puzzle.itc.mobiliar.common.util.ApplicationServerContainer.APPSERVERCONTAINER;
-import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
-import static javax.ws.rs.core.Response.Status.CREATED;
-import static javax.ws.rs.core.Response.Status.NOT_FOUND;
-import static javax.ws.rs.core.Response.Status.OK;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import javax.ws.rs.core.Response;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-
 import ch.mobi.itc.mobiliar.rest.dtos.ResourceGroupDTO;
 import ch.mobi.itc.mobiliar.rest.dtos.ResourceReleaseCopyDTO;
 import ch.mobi.itc.mobiliar.rest.dtos.ResourceReleaseDTO;
 import ch.mobi.itc.mobiliar.rest.exceptions.ExceptionDto;
 import ch.puzzle.itc.mobiliar.business.deploy.boundary.DeploymentBoundary;
 import ch.puzzle.itc.mobiliar.business.environment.entity.ContextEntity;
-import ch.puzzle.itc.mobiliar.business.generator.control.extracted.ResourceDependencyResolverService;
 import ch.puzzle.itc.mobiliar.business.releasing.boundary.ReleaseLocator;
 import ch.puzzle.itc.mobiliar.business.releasing.entity.ReleaseEntity;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.boundary.CopyResource;
@@ -64,12 +37,29 @@ import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.Resource;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceEntity;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceGroupEntity;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceTypeEntity;
-import ch.puzzle.itc.mobiliar.business.security.boundary.PermissionBoundary;
 import ch.puzzle.itc.mobiliar.business.server.boundary.ServerView;
 import ch.puzzle.itc.mobiliar.common.exception.AMWException;
 import ch.puzzle.itc.mobiliar.common.exception.ElementAlreadyExistsException;
 import ch.puzzle.itc.mobiliar.common.exception.NotFoundException;
 import ch.puzzle.itc.mobiliar.common.exception.ValidationException;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import javax.ws.rs.core.Response;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+import static ch.puzzle.itc.mobiliar.common.util.ApplicationServerContainer.APPSERVERCONTAINER;
+import static javax.ws.rs.core.Response.Status.*;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 public class ResourceGroupsRestTest {
@@ -106,12 +96,6 @@ public class ResourceGroupsRestTest {
 
     @Mock
     DeploymentBoundary deploymentBoundaryMock;
-
-    @Mock
-    ResourceDependencyResolverService resourceDependencyResolverServiceMock;
-
-    @Mock
-    PermissionBoundary permissionBoundaryMock;
 
     @Test
     public void getResourcesWhenTypeIsNullAndNoRessourceGroupsShouldReturnEmptyResult() {
@@ -437,71 +421,6 @@ public class ResourceGroupsRestTest {
         assertThat(((ExceptionDto) response.getEntity()).getMessage(), is("Resource not found"));
     }
 
-    @Test
-    public void shouldReturnNotFoundWhenGetCopyFromCandidatesForNonExistingResource() {
-        // given
-        when(resourceLocatorMock.getResourceById(999)).thenReturn(null);
 
-        // when
-        Response response = rest.getCopyFromCandidates(999);
-
-        // then
-        assertThat(response.getStatus(), is(NOT_FOUND.getStatusCode()));
-    }
-
-    @Test
-    public void shouldReturnForbiddenWhenNoCopyPermission() {
-        // given
-        ResourceEntity resource = mock(ResourceEntity.class);
-        when(resourceLocatorMock.getResourceById(1)).thenReturn(resource);
-        when(permissionBoundaryMock.canCopyFromResource(resource)).thenReturn(false);
-
-        // when
-        Response response = rest.getCopyFromCandidates(1);
-
-        // then
-        assertThat(response.getStatus(), is(Response.Status.FORBIDDEN.getStatusCode()));
-    }
-
-    @Test
-    public void shouldReturnOkOnSuccessfulCopyFromResourceById() throws AMWException {
-        // given
-        CopyResourceResult copyResourceResult = mock(CopyResourceResult.class);
-        when(copyResourceResult.isSuccess()).thenReturn(true);
-        when(copyResourceMock.doCopyResource(1, 2)).thenReturn(copyResourceResult);
-
-        // when
-        Response response = rest.copyFromResourceById(1, 2);
-
-        // then
-        assertThat(response.getStatus(), is(OK.getStatusCode()));
-    }
-
-    @Test
-    public void shouldReturnBadRequestWhenCopyFromResourceByIdFails() throws AMWException {
-        // given
-        CopyResourceResult copyResourceResult = mock(CopyResourceResult.class);
-        when(copyResourceResult.isSuccess()).thenReturn(false);
-        when(copyResourceMock.doCopyResource(1, 2)).thenReturn(copyResourceResult);
-
-        // when
-        Response response = rest.copyFromResourceById(1, 2);
-
-        // then
-        assertThat(response.getStatus(), is(BAD_REQUEST.getStatusCode()));
-    }
-
-    @Test
-    public void shouldReturnBadRequestWhenCopyFromResourceByIdThrowsException() throws AMWException {
-        // given
-        when(copyResourceMock.doCopyResource(1, 2)).thenThrow(new AMWException("Permission Denied"));
-
-        // when
-        Response response = rest.copyFromResourceById(1, 2);
-
-        // then
-        assertThat(response.getStatus(), is(BAD_REQUEST.getStatusCode()));
-        assertThat(((ExceptionDto) response.getEntity()).getMessage(), is("Permission Denied"));
-    }
 
 }
